+++ analytics/analytics-listeners/examples/helpers/index.tsx
@@ -92,6 +92,17 @@ export class DummyPeopleTeamsComponent extends React.Component<Props> {
   }
 }
 
+export class DummyNotificationsComponent extends React.Component<Props> {
+  render() {
+    return (
+      <CustomButton
+        text={FabricChannel.notifications}
+        onClick={this.props.onClick}
+      />
+    );
+  }
+}
+
 class MyButton extends React.Component<Props> {
   static displayName = 'MyButton';
   render() {
@@ -110,6 +121,7 @@ const componentChannels = {
   [FabricChannel.editor]: DummyEditorComponent,
   [FabricChannel.media]: DummyMediaComponent,
   [FabricChannel.peopleTeams]: DummyPeopleTeamsComponent,
+  [FabricChannel.notifications]: DummyNotificationsComponent,
 };
 
 export const createComponentWithAnalytics = (channel: FabricChannel) =>