+++ design-system/menu/src/menu-item/__tests__/visual-regression/link.tsx
@@ -9,6 +9,8 @@ import {
 
 const linkItem = '[data-testid="item-link"]';
 const linkItemSelected = '[data-testid="item-link-selected"]';
+const linkItemMultipleLine = '[data-testid="item-link-multiple-line"]';
+const linkItemLongUrl = '[data-testid="item-link-long-url-multiple-line"]';
 
 const url = getExampleUrl(
   'design-system',
@@ -38,5 +40,13 @@ describe('<LinkItem />', () => {
 
   it('should match selected item', async () => {
     await verifyElementMatchProductionImage(linkItemSelected);
+  });
+
+  it('should match item with lots of text - wrap text', async () => {
+    await verifyElementMatchProductionImage(linkItemMultipleLine);
+  });
+
+  it('should match item with lots of text - wrap text for long url', async () => {
+    await verifyElementMatchProductionImage(linkItemLongUrl);
   });
 });