+++ analytics/analytics-namespaced-context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/analytics-namespaced-context
 
+## 6.1.0
+
+### Minor Changes
+
+- [`1ecfe397c00`](https://bitbucket.org/atlassian/atlassian-frontend/commits/1ecfe397c00) - Added a notification channel
+
 ## 6.0.0
 
 ### Major Changes